KAFKA-18847: Refactor OAuth layer to improve reusability
