+++ clients/src/main/java/org/apache/kafka/common/security/oauthbearer/internals/secured/AccessTokenRetrieverFactory.java
@@ -75,7 +75,7 @@ public class AccessTokenRetrieverFactory  {
 
             boolean urlencodeHeader = validateUrlencodeHeader(cu);
 
-            return new HttpAccessTokenRetriever(clientId,
+            return new ClientCredentialsAccessTokenRetriever(clientId,
                 clientSecret,
                 scope,
                 sslSocketFactory,
